Implement sorting of metric groups, based on a number of attributes.
This implements "in-memory" sorting of metric groups, using a number of
different attributes. As we don't currently do pagination or have any
performance concerns it was simpler to load all the data and sort it
locally, rather than adding sorting as a capability to the API.

As there are a number of different attributes to sort on, it allows
configuration of these attributes based on a "keypath", which is an
array of symbols, which name the methods to be called to fetch the value
to sort on.

It also simplifies the "MetricPresenter" classes by moving the fetching
of the root entities into the controller, instead of in an over-loaded
initializer in each of the sub-classes.