Add Javascript based feature test.
This commit adds Javascript-enabled testing, driving headless Chrome
with Selenium. It uses a `with_conditional_javascript` to run the test
twice - once with Javascript enabled and once with it disabled.

The tests have to be modified slighty (the non-Javascript one has to
click the 'Apply' button, for example), but I feel like the slight
additional complextity outweighs the duplication of the main test logic.