Don't show metrics and items that are N/A
If a metric item is N/A then we don't show it at all.  If none of the
items in a metric are applicable then we don't show that column, showing
text instead.

Currently needs more text, and fixing of the html so that when
filtering by a metric then the item is collapsed.