Don't show metrics and items that are not provided
If a metric item is not provided we show a label.  If none of the
items in a metric are provided then we don't show that column, showing
text instead.

Currently needs more text, and fixing of the html so that when
filtering by a metric then the item is collapsed.